moved back to single thread fro now, need to fix some performance issues and rethink clearly how to use parallelism

+++ src/audio_file/analysis/clipping_samples.rs
@@ -2,7 +2,7 @@ use crate::audio_file::Signal;
 
 impl super::ClippingSamples {
     #[inline]
-    pub fn process(samples: &Signal) -> usize {
+    pub fn process(samples: &Signal) -> u64 {
         samples
             .iter()
             .filter(|&&x| is_clipping(x))
